Add more support for gridxc discovery and enable WITH_GRID_SP
The search_for_gridxc.cmake logic has been enhanced to be able
to discover auto-tools-compiled versions of libgridxc, both with
and without the 'multiconfig' feature.

In all cases, the found library is checked for compatibility with the
current options (MPI, LIBXC, GRID_SP), and skipped if not compatible.
Fallbacks to compilation of code in submodules or 'fetched' are still
available.

The global option WITH_GRID_SP (default OFF) has been added to treat
grid arrays in single precision.