enh: Add CMake support for Doxygen and FORD documentation (wip)
Added CMake scaffolding to create the Doxygen pages, and also (when an
appropriate version of FORD is used), the FORD pages.  Both are
developer-oriented, and incomplete as of now.

Some markdown pages from the FORD side of the documentation have links
that cannot be resolved, so the main page for Doxygen is just the main
README.md file, which has been updated to reference the new Siesta
documentation site.

To enable Doxygen documentation, enable BUILD_DOCS.
To enable FORD documentation, enable BUILD_DOCS_FORD.

(Inspiration from SIRIUS and DBCSR projects)