Implement a simpler interface with wannier90
The main change is the wrapping of the ``wannier_prog`` program to
turn it into a subroutine that can be called directly from another
program. This is now the only entry point of the API, as implemented in
the ``wannier90_m`` module.

In addition, a simple CMake building system has been implemented to facilitate
the automatic compilation of the wrapper.

The ``wannier90_wrapper`` subroutine accepts as arguments:

* The ``seedname`` or file prefix.
* The MPI communicator to be used (only in the MPI version)
* (Optional) nnkp_mode: a flag to request only the generation of a .nnkp file.
* (Optional) dryrun_mode: a flag to perform a simple dry-run to check the input file.
* (Optional) Arrays to hold the information about k-point neighbors
* (Optional) Arrays to hold the information about the unitary matrices

The last two are convenience arguments to enable client codes to extract the needed information
directly, without having to read the .nnkp and .chk files, respectively.

All other interaction with wannier90 is through files:

* ``seedname.win`` (provided by client) is the input file
* ``seedname.amn`` (provided by client)
* ``seedname.mmn`` (provided by client)
* ``seedname.{eig_ext}`` (provided by client)
* ``seedname.wout`` (generated by wannier90)
*  ... plus other plotting and auxiliary files generated by wannier90.

The standard command-line interaction with wannier90 has been bypassed.