Turn units into protected variables, and initialize using the units table
Some changes have to be made in a number of places. In particular, a
'units_initialization' call has to be made before the units can be used.

(This is in addition to the setting of the unit handler for fdf. Sometimes
the two calls can be combined; sometimes they cannot).