large rework of the cmake infrastructure
This should alleviate some of the decisions that users
needs to make.
For instance instead of forcing users to take a stance
on every single library, we instead enable the library IFF
it is accessible and usable.
There might be certain libraries that are not defaulted
to be compiled in, in which case the default should not be
TRUE. However, in those cases we should clearly spell
out that the users needs to take a decision.

Beautification of the output to a great extent.
Many places I have added an indentation to make grouping
a little softer on the reader. This could be expanded further.

The search_*.cmake files have been moved to FindCustom*.cmake
to make use of find_package. This will allow one to test whether
a package exists, and then make a decision based on whether it does or not.

Added the BLAS library in case LAPACK does not add it.
BLAS is not required if LAPACK has the functionality required.

Signed-off-by: Nick Papior <nickpapior@gmail.com>