Adding build-info print-outs
Not all packages are there, but a large fraction.
The print-out is done via a function, so there is some
limitation. We could expand this as we go by.
The nice thing about the function is that it will be
consistent for all the library features.

I would prefer some kind of grouping based on

- REQUIRED (Libgridxc ...)
- RECOMMENDED (LIBXC ...)
- ADVANCED (OpenMP ...)

Lets see.

I have changed WITH_LUA to WITH_FLOOK as that
should be more natural WITH_LUA and FLOOK_ROOT
was necessary. Now they are streamlined.

Signed-off-by: Nick Papior <nickpapior@gmail.com>