Fixed implicitly linked BLAS,LAPACK,ScaLAPACK + info
BLAS,LAPACK and ScaLAPACK may be implicitly linked.
This commit should fix this by setting the libraries
to NONE when it is originally empty but found.

Also made the print-info block things together.
It is much simpler to read for a user, albeit
the output is much longer.

Circumvented a case when variables got defined, with
no value. Now we check whether variables are empty
to print them in the correct place.

Changed scalapack target to SCALAPACK::SCALAPACK.
The other linear algebra packages are named
like that, so might be simpler to follow the same
pattern.

Signed-off-by: Nick Papior <nickpapior@gmail.com>