Fixed rebase problems and fixed cmake builds
Now the units has been moved to a separate directory
to decouple it from the rest of siesta.
This method could be useful in other places.

The units are now created in a units.py script
and this is able to populate the tables in the units_*.f90
modules.
Their common usage functions are diverted into the units_common
module which also holds the inquire_unit function.

I am able to build everything with cmake.

TODO add the simple makefile scheme.

Signed-off-by: Nick Papior <nickpapior@gmail.com>