Added more arguments to Siesta_find_package
The CMAKE and PKG_CONFIG variables now determine
the searched package name in the respective package
finders. These are by default the package name.
This forces users to be explicit at invocation time,
which shouldn't be a problem since it clarifies intent.
The outcome of the Siesta_find_package still returns
in the <NAME>_FOUND and <UPPER:NAME>_FOUND variables.

A fix for the find_package(... REQUIRED) is now done.
It requires Siesta_find_package to check the <NAME>_FIND_REQUIRED
variable.

The same happens for the find_package(... QUIET) which
uses the <NAME>_FIND_QUIETLY variable. This is now
obeyed by not using an explicit argument to Siesta_find_package.

Fixed variable naming for the GIT_REPOSITORY argument.

Fixed a bug where the tag and repo-url was not caught.

Signed-off-by: Nick Papior <nickpapior@gmail.com>