Completed libgridxc migration to Siesta_find_package
This now is much simpler to follow.

It also removes the possibility for the cmake stuff
to create double targets if the same library is found
multiple times.
The current way to disable this is to explicitly
set the LIBGRIDXC_FIND_METHOD=...
to disable which ones are faulty.

The checks for libgridxc required some stub functions
which are now added to the compilation checks.
This allows one to use shared libraries and now works
correctly. Yay.

Signed-off-by: Nick Papior <nickpapior@gmail.com>