Fixed linker errors for static libraries
There are still issues in the linker order.
I don't know why it doesn't work.
See remark in Src/CMakeLists.txt.

Added quotation around the generator tools in Src/MPI.

Trying to fix how the CMAKE_BUILD_TYPE works.
It should now work *after* toolchains are set
such that toolchains can denote the build-type.

libpsml and libgridxc are now linkable with
the older versions without procedure pointers.
This seems relevant since it does not force
the entire software stack to be bumped (e.g.
bundles etc.).

Some formatting fixes in the cmake files.

Restructured a bit the SCALAPACK library which
should make it more stable for dependencies.

Fixed lots of paths omitting "". Essentially
*all* variable references that are paths
should be wrapped in "..." to ensure they work
with spaces. Still Siesta fails to compile
with spaces in the path, but this could also be
because of sub-projects.

Added more print-out such that developers can
get information related to all libraries.
simply do:
  cmake --log-level=debug
and the printinfo blocks will be pasted with lots more information
for debugging libraries and dependencies.

Signed-off-by: Nick Papior <nickpapior@gmail.com>