Fix compilation of utilities when libgridxc does not have procedure pointers
A recent commit introduced code to deal with versions of libgridxc and libpsm
without procedure pointers.

The gridxc handlers 'alloc_memory_event' and 'alloc_error_report' were not
resolved in the compilation of programs in Util. (They were for Siesta, with
special versions at the end of 'alloc_handlers.F90'.)

The easiest fix (short of having to re-introduce different 'handler'
files for utilities) is to put handlers in the 'external interfaces' section
of Src/libsys.

This has the side-effect of exposing duplicate symbols for the
compilation of Siesta. For expediency, the versions in
'alloc_handlers.F90' have been deactivated, with the consequence that
Siesta cannot report allocations/deallocations coming from inside
libgridxc.

This *could* be fixed with different libsys versions for Siesta and for
the utilities.